Add more compiler flags
- Include the '-Wshadow' and '-Wconversion' compiler flags, which are
  not enabled by '-Wall' and '-Wextra';

- Fix compilation errors in files which were affected by the addition
  of the two new flags above, while trying to compile with gcc, g++
  and clang compilers.

+++ test/rand-perm/rand-perm.c
@@ -27,7 +27,7 @@ void rand_perm_gen_init(rand_perm_gen_t *gen, int m, int n)
 
     int c = n - m + 1;
 
-    gen->nums = (int *)calloc(c, sizeof(int));
+    gen->nums = (int *)calloc((size_t)c, sizeof(int));
 
     if (NULL == gen->nums)
         return;
@@ -75,7 +75,7 @@ static int rand_between(int m, int n)
     static bool init_seed = false;
 
     if (!init_seed) {
-        srand(time(NULL));
+        srand((unsigned int)time(NULL));
         init_seed = true;
     }
